Moved CompImageHeader to separate file
